Culminacion del Proyecto

+++ src/app/doctores/doctores-list/doctores-list.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Doctores } from '../../interfaces/doctores';
 import { DoctoresService } from '../../services/doctores.service';
 import swal from 'sweetalert2';
@@ -6,9 +6,9 @@ import swal from 'sweetalert2';
 @Component({
   selector: 'app-doctores-list',
   templateUrl: './doctores-list.component.html',
-  styleUrl: './doctores-list.component.css'
+  styleUrls: ['./doctores-list.component.css']
 })
-export class DoctoresListComponent {
+export class DoctoresListComponent implements OnInit {
 
   doctores: Doctores[] = [];
 
@@ -25,10 +25,25 @@ export class DoctoresListComponent {
   }
 
   deleteCustomer(id: number): void {
-    this.doctoresService.deleteDoctor(id).subscribe(() => {
-      this.getDoctores(); // Refresca la lista de clientes después de eliminar
+    swal.fire({
+      title: '¿Estás seguro?',
+      text: 'No podrás revertir esto',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.doctoresService.deleteDoctor(id).subscribe(() => {
+          this.getDoctores(); // Refresca la lista de doctores después de eliminar
+          swal.fire(
+            'Eliminado',
+            'El doctor ha sido eliminado',
+            'success'
+          );
+        });
+      }
     });
   }
-
-
 }
